chore: move crops out of items.js
  this breaks up items.js into more files.
  - adds crop.js
  - adds crops directory for all crops to have a file in, plus an index
