+++ src/data/crops/index.js
@@ -1,0 +1,11 @@
+export { asparagus, asparagusSeed } from './carrot'
+export { carrot, carrotSeed } from './carrot'
+export { corn, cornSeed } from './corn'
+export { jalapeno, jalapenoSeed } from './corn'
+export { onion, onionSeed } from './onion'
+export { potato, potatoSeed } from './potato'
+export { pumpkin, pumpkinSeed } from './pumpkin'
+export { soybean, soybeanSeed } from './soybean'
+export { spinach, spinachSeed } from './spinach'
+export { tomato, tomatoSeed } from './tomato'
+export { wheat, wheatSeed } from './wheat'
