+++ src/data/items.js
@@ -2,7 +2,7 @@
  * @module farmhand.items
  */
 
-import { cropLifeStage, cropType, fieldMode, itemType } from '../enums'
+import { fieldMode, itemType } from '../enums'
 import {
   COW_FEED_ITEM_ID,
   HUGGING_MACHINE_ITEM_ID,
@@ -11,22 +11,7 @@ import {
 
 const { freeze } = Object
 const {
-  ASPARAGUS,
-  CARROT,
-  CORN,
-  JALAPENO,
-  ONION,
-  POTATO,
-  PUMPKIN,
-  SOYBEAN,
-  SPINACH,
-  TOMATO,
-  WHEAT,
-} = cropType
-const { SEED, GROWING } = cropLifeStage
-const {
   COW_FEED,
-  CROP,
   FERTILIZER,
   HUGGING_MACHINE,
   MILK,
@@ -34,322 +19,30 @@ const {
   SPRINKLER,
 } = itemType
 
-////////////////////////////////////////
-//
-// CROPS
-//
-////////////////////////////////////////
-
-const crop = ({
-  cropTimetable,
-  growsInto,
-  tier,
-
-  isSeed = Boolean(growsInto),
-  cropLifecycleDuration = Object.values(cropTimetable).reduce(
-    (acc, value) => acc + value,
-    0
-  ),
-
-  ...rest
-}) =>
-  freeze({
-    cropTimetable,
-    doesPriceFluctuate: true,
-    tier,
-    type: CROP,
-    value: 10 + cropLifecycleDuration * tier * (isSeed ? 1 : 3),
-    ...(isSeed && {
-      enablesFieldMode: fieldMode.PLANT,
-      growsInto,
-      isPlantableCrop: true,
-    }),
-    ...rest,
-  })
-
-const fromSeed = ({ cropTimetable, cropType, growsInto, tier }) => ({
-  cropTimetable,
-  cropType,
-  doesPriceFluctuate: true,
-  id: growsInto,
-  tier,
-  type: CROP,
-})
-
-/**
- * @property farmhand.module:items.carrotSeed
- * @type {farmhand.item}
- */
-export const carrotSeed = crop({
-  cropType: CARROT,
-  cropTimetable: {
-    [SEED]: 2,
-    [GROWING]: 3,
-  },
-  growsInto: 'carrot',
-  id: 'carrot-seed',
-  name: 'Carrot Seed',
-  tier: 1,
-})
-
-/**
- * @property farmhand.module:items.carrot
- * @type {farmhand.item}
- */
-export const carrot = crop({
-  ...fromSeed(carrotSeed),
-  name: 'Carrot',
-})
-
-/**
- * @property farmhand.module:items.asparagusSeed
- * @type {farmhand.item}
- */
-export const asparagusSeed = crop({
-  cropType: ASPARAGUS,
-  cropTimetable: {
-    [SEED]: 4,
-    [GROWING]: 5,
-  },
-  growsInto: 'asparagus',
-  id: 'asparagus-seed',
-  name: 'Asparagus Seed',
-  tier: 4,
-})
-
-/**
- * @property farmhand.module:items.asparagus
- * @type {farmhand.item}
- */
-export const asparagus = crop({
-  ...fromSeed(asparagusSeed),
-  name: 'Asparagus',
-})
-
-/**
- * @property farmhand.module:items.pumpkinSeed
- * @type {farmhand.item}
- */
-export const pumpkinSeed = crop({
-  cropType: PUMPKIN,
-  cropTimetable: {
-    [SEED]: 3,
-    [GROWING]: 5,
-  },
-  growsInto: 'pumpkin',
-  id: 'pumpkin-seed',
-  name: 'Pumpkin Seed',
-  tier: 1,
-})
-
-/**
- * @property farmhand.module:items.pumpkin
- * @type {farmhand.item}
- */
-export const pumpkin = crop({
-  ...fromSeed(pumpkinSeed),
-  name: 'Pumpkin',
-})
-
-/**
- * @property farmhand.module:items.spinachSeed
- * @type {farmhand.item}
- */
-export const spinachSeed = crop({
-  cropType: SPINACH,
-  cropTimetable: {
-    [SEED]: 2,
-    [GROWING]: 4,
-  },
-  growsInto: 'spinach',
-  id: 'spinach-seed',
-  name: 'Spinach Seed',
-  tier: 1,
-})
-
-/**
- * @property farmhand.module:items.spinach
- * @type {farmhand.item}
- */
-export const spinach = crop({
-  ...fromSeed(spinachSeed),
-  name: 'Spinach',
-})
-
-/**
- * @property farmhand.module:items.cornSeed
- * @type {farmhand.item}
- */
-export const cornSeed = crop({
-  cropType: CORN,
-  cropTimetable: {
-    [SEED]: 3,
-    [GROWING]: 7,
-  },
-  growsInto: 'corn',
-  id: 'corn-seed',
-  name: 'Corn Kernels',
-  tier: 2,
-})
-
-/**
- * @property farmhand.module:items.corn
- * @type {farmhand.item}
- */
-export const corn = crop({
-  ...fromSeed(cornSeed),
-  name: 'Corn',
-})
-
-/**
- * @property farmhand.module:items.potatoSeed
- * @type {farmhand.item}
- */
-export const potatoSeed = crop({
-  cropType: POTATO,
-  cropTimetable: {
-    [SEED]: 2,
-    [GROWING]: 3,
-  },
-  growsInto: 'potato',
-  id: 'potato-seed',
-  name: 'Potato Seeds',
-  tier: 2,
-})
-
-/**
- * @property farmhand.module:items.potato
- * @type {farmhand.item}
- */
-export const potato = crop({
-  ...fromSeed(potatoSeed),
-  name: 'Potato',
-})
-
-/**
- * @property farmhand.module:items.onionSeed
- * @type {farmhand.item}
- */
-export const onionSeed = crop({
-  cropType: ONION,
-  cropTimetable: {
-    [SEED]: 3,
-    [GROWING]: 4,
-  },
-  growsInto: 'onion',
-  id: 'onion-seed',
-  name: 'Onion Seeds',
-  tier: 2,
-})
-
-/**
- * @property farmhand.module:items.onion
- * @type {farmhand.item}
- */
-export const onion = crop({
-  ...fromSeed(onionSeed),
-  name: 'Onion',
-})
-
-/**
- * @property farmhand.module:items.soybeanSeed
- * @type {farmhand.item}
- */
-export const soybeanSeed = crop({
-  cropType: SOYBEAN,
-  cropTimetable: {
-    [SEED]: 2,
-    [GROWING]: 2,
-  },
-  growsInto: 'soybean',
-  id: 'soybean-seed',
-  name: 'Soybean Seeds',
-  tier: 3,
-})
-
-/**
- * @property farmhand.module:items.soybean
- * @type {farmhand.item}
- */
-export const soybean = crop({
-  ...fromSeed(soybeanSeed),
-  name: 'Soybean',
-})
-
-/**
- * @property farmhand.module:items.wheatSeed
- * @type {farmhand.item}
- */
-export const wheatSeed = crop({
-  cropType: WHEAT,
-  cropTimetable: {
-    [SEED]: 1,
-    [GROWING]: 1,
-  },
-  growsInto: 'wheat',
-  id: 'wheat-seed',
-  name: 'Wheat Seeds',
-  tier: 3,
-})
-
-/**
- * @property farmhand.module:items.wheat
- * @type {farmhand.item}
- */
-export const wheat = crop({
-  ...fromSeed(wheatSeed),
-  name: 'Wheat',
-})
-
-/**
- * @property farmhand.module:items.tomatoSeed
- * @type {farmhand.item}
- */
-export const tomatoSeed = crop({
-  cropType: TOMATO,
-  cropTimetable: {
-    [SEED]: 5,
-    [GROWING]: 6,
-  },
-  growsInto: 'tomato',
-  id: 'tomato-seed',
-  name: 'Tomato Seeds',
-  tier: 3,
-})
-
-/**
- * @property farmhand.module:items.tomato
- * @type {farmhand.item}
- */
-export const tomato = crop({
-  ...fromSeed(tomatoSeed),
-  name: 'Tomato',
-})
-
-/**
- * @property farmhand.module:items.jalapenoSeed
- * @type {farmhand.item}
- */
-export const jalapenoSeed = crop({
-  cropType: JALAPENO,
-  cropTimetable: {
-    [SEED]: 2,
-    [GROWING]: 3,
-  },
-  growsInto: 'jalapeno',
-  id: 'jalapeno-seed',
-  name: 'Jalapeño Seed',
-  tier: 4,
-})
-
-/**
- * @property farmhand.module:items.jalapeno
- * @type {farmhand.item}
- */
-export const jalapeno = crop({
-  ...fromSeed(jalapenoSeed),
-  name: 'Jalapeño',
-})
+export {
+  asparagus,
+  asparagusSeed,
+  carrot,
+  carrotSeed,
+  corn,
+  cornSeed,
+  jalapeno,
+  jalapenoSeed,
+  onion,
+  onionSeed,
+  potato,
+  potatoSeed,
+  pumpkin,
+  pumpkinSeed,
+  soybean,
+  soybeanSeed,
+  spinach,
+  spinachSeed,
+  tomato,
+  tomatoSeed,
+  wheat,
+  wheatSeed,
+} from './crops'
 
 ////////////////////////////////////////
 //
